% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestNetR.R
\name{create_breeding_data_list}
\alias{create_breeding_data_list}
\title{Create list of breeding-period segments}
\usage{
create_breeding_data_list(dir.raw, auto = TRUE, segment_days = 1)
}
\arguments{
\item{dir.raw}{Character; directory containing raw \code{.lux} and \code{.deg} files
(e.g. \code{"data/RawData/SpeciesName/"}).}

\item{auto}{Logical; if \code{TRUE} (default), breeding periods are detected
automatically using \code{set_breeding_period()}. Otherwise, manually annotation is required.a}
}
\value{
A named list of individuals, each containing its own list of
breeding-period segments as lists of vectors.
}
\description{
Reads individual \code{.lux} and \code{.deg} files from a directory, determines each
bird’s breeding period, preprocesses the data, and stores the resulting
segmented time series per individual. Each list element corresponds to one
individual ID and contains a list of its breeding-period segments.
}
