% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestNetR.R
\name{preprocessing}
\alias{preprocessing}
\title{Assemble and segment biologger data for the breeding period}
\usage{
preprocessing(
  ID,
  raw_light,
  raw_deg,
  tm.breeding,
  dir.breeding,
  tz = "UTC",
  segment_days = 1,
  overlap_days = 1,
  Light_quantiles = c(0.9478656, 11.2115639),
  Tmin_quantiles = c(-2.8, 23.5),
  Tmax_quantiles = c(3, 40.8)
)
}
\arguments{
\item{ID}{Character scalar; individual or deployment ID.}

\item{raw_light}{data.frame with POSIXct-like column \code{Date} (from \code{read_lux()}).}

\item{raw_deg}{data.frame with POSIXct-like column \code{Date} (from \code{read_deg()}).}

\item{tm.breeding}{Named vector of class \code{Date} (length 2; names 'start','end').}

\item{dir.breeding}{Character scalar; directory to save breeding data as .csv.}

\item{tz}{Character scalar, time zone (default "UTC").}

\item{segment_days}{integer, segment length in days (default 2).}

\item{overlap_days}{integer, overlap in days (default 1).}

\item{Light_quantiles}{numeric length 2, rescaling bounds.}

\item{Tmin_quantiles}{numeric length 2, rescaling bounds.}

\item{Tmax_quantiles}{numeric length 2, rescaling bounds.}
}
\value{
A list with raw data, segmented windows, and flattened training data.
}
\description{
Cleans light/temperature data, subsets to breeding window, creates overlapping segments,
rescales values based on pre-defined quantiles, and flattens them into training-ready lists.
}
